Dior_assignment_project/vestiaire_scraper.py
import re
import asyncio
from datetime import datetime
from playwright.async_api import async_playwright
from bs4 import BeautifulSoup
import pandas as pd

async def scrape_vestiaire_for_product(product_row):
    """Cherche un produit spécifique sur Vestiaire Collective."""
    # Ta logique de regex pour nettoyer le nom du produit
    keywords = re.sub(r'dior|christian|sac |handbag |pochette ', '', product_row['product_name'], flags=re.IGNORECASE).strip()
    search_query = f"Dior {keywords}"
    
    # ... (Copie ici ta logique de scraping Vestiaire Collective) ...
    # N'oublie pas de retourner une liste de dictionnaires