Dior_assignment_project/requirements.txt
google-cloud-bigquery
pandas
pyarrow
playwright
beautifulsoup4
nest_asyncio
jupyterlab